Separate lock paths
Prior to this PR the same ZNode (path) was used by the `InterProcessSemaphoreMutex` lock and the node lock. The `InterProcessSemaphoreMutex` lock is a curator recipe. Based on the [Curator Tech Note 7] it is not allowed to use the paths passed to Curator recipes. Therefore, both locks using now a different ZNode.